DOMA-2461 Some tests were added
